feat: implement asynq tasks for closing votings (selection, rating), implement session repo, implement extra middlewares

+++ internal/model/poll_option.go
@@ -8,6 +8,6 @@ type PollOption struct {
 	PollID      int64 `gorm:"not null"`
 	Poll        Poll  `gorm:"foreignKey:PollID"`
 	OptionIndex int   `gorm:"not null"` // 0, 1, 2, 3... (position in poll)
-	MovieID     int64 `gorm:"not null"`
+	MovieID     int   `gorm:"not null"`
 	Movie       Movie `gorm:"foreignKey:MovieID"`
 }
